test(talk): split WebRTC tool lifecycle coverage
